+++ docs/settings.jl
@@ -11,8 +11,12 @@ settings = Dict(
     "Getting Started with ITensor" => [
       "Installing Julia and ITensor" => "getting_started/Installing.md",
       "Running ITensor and Julia Codes" => "getting_started/RunningCodes.md",
-      "Tutorials" => "getting_started/Tutorials.md",
       "Next Steps" => "getting_started/NextSteps.md",
+    ],
+    "Tutorials" => [
+      "DMRG" => "tutorials/DMRG.md",
+      "Quantum Number Conserving DMRG" => "tutorials/QN_DMRG.md",
+      "MPS Time Evolution" => "tutorials/MPSTimeEvolution.md",
     ],
     "Code Examples" => [
       "ITensor Examples" => "examples/ITensor.md",
@@ -42,7 +46,7 @@ settings = Dict(
     "Advanced usage guide" => [
       "Advanced usage guide" => "AdvancedUsageGuide.md",
       "Multithreading" => "Multithreading.md",
-      "Symmetric tensor background and usage" => "QNTricks.md",
+      "Symmetric (QN conserving) tensors: background and usage" => "QNTricks.md",
       "Timing and profiling" => "CodeTiming.md",
       "Contraction sequence optimization" => "ContractionSequenceOptimization.md",
     ],