+++ src/mps/abstractmps.jl
@@ -42,7 +42,7 @@ end
 The element type of the MPS/MPO. Always returns `ITensor`.
 
 For the element type of the ITensors of the MPS/MPO,
-use `common_itensor_eltype`.
+use `promote_itensor_eltype`.
 """
 eltype(::AbstractMPS) = ITensor
 
@@ -1495,7 +1495,7 @@ function setindex!(
     sites = sites0[[perm...]]
     # Check if the site indices
     # are fermionic
-    if any(anyfermionic, sites)
+    if !using_auto_fermion() && any(anyfermionic, sites)
       if length(sites) == 2 && ψ isa MPS
         if all(allfermionic, sites)
           s0 = Index.(sites0)