+++ test/runtests.jl
@@ -41,6 +41,7 @@ using Test
     "qnitensor.jl",
     "qncombiner.jl",
     "qndiagitensor.jl",
+    "fermions.jl",
     "empty.jl",
     "qnmpo.jl",
     "readwrite.jl",